% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amperometry2df.R
\name{amperometry2df}
\alias{amperometry2df}
\title{Read amperometry data from CHI760}
\usage{
amperometry2df(datafilename, wearea = 1)
}
\arguments{
\item{datafilename}{text string with full path to experimental file}

\item{wearea}{(optional) area of working electrode (in square centimeter)}
}
\value{
Dataframe with the following columns (and no extra attributes):
   $ sampleid        : chr
   $ time            : num [seconds]
   $ current         : num [ampere]
   $ currentdensity  : num [ampere per square cm]
   $ timediff        : num [seconds]
   $ dIdt            : num
   $ didt            : num
   $ charge          : num [coulomb]
   $ chargedensity   : num [coulomb per square cm]
   $ InitE           : num
   $ SampleInterval  : num
   $ RunTime         : num
   $ QuietTime       : num
   $ Sensitivity     : num
}
\description{
Reads current-time data (from CHI 760 potentiostat)
and returns a dataframe with the data,
the data attributes (experimental conditions),
and some calculated parameters (charge, didt, etc.)
}
\details{
The CH Instruments 760 potentiostat records all data using
   standard SI units, therefore this function assumes all potentials
   to be in volts, currents to be in amperes, charges in Coulombs,
   time in seconds, and so on.
}

