% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronoamp2df.R
\name{chronoamp2df}
\alias{chronoamp2df}
\title{Read chronoamperometry data from CHI760}
\usage{
chronoamp2df(datafilename, wearea = 1)
}
\arguments{
\item{datafilename}{text string with full path to experimental file}

\item{wearea}{(optional) area of working electrode (in square centimeter)}
}
\value{
Dataframe with the following columns (and no extra attributes):
   $ sampleid        : chr
   $ step            : num
   $ time            : num
   $ current         : num
   $ currentdensity  : num
   $ InitE           : num
   $ HighE           : num
   $ LowE            : num
   $ InitPN          : chr
   $ Step            : num
   $ Pulse width     : num
   $ Sample interval : num
   $ Quiet Time      : num
   $ Sensitivity     : num
}
\description{
Reads current-time data (from CHI 760 potentiostat)
and returns a dataframe with the data and the data attributes
(experimental conditions),
}
\details{
The CH Instruments 760 potentiostat records all data using
   standard SI units, therefore this function assumes all potentials
   to be in volts, currents to be in amperes, charges in Coulombs,
   time in seconds, and so on.
}

