% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv2df.R
\name{cv2df}
\alias{cv2df}
\title{Read CV data from CHI760}
\usage{
cv2df(datafilename, wearea = 1)
}
\arguments{
\item{datafilename}{path to file containing data}

\item{wearea}{working electrode's area in sqcm}
}
\value{
dataframe with data
}
\description{
Read CV data from CHI760
}
\details{
The CH Instruments 760 potentiostat records all data using
   standard SI units, therefore this function assumes all potentials
   to be in volts, currents to be in amperes, charges in Coulombs,
   time in seconds, and so on.
}

