% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp2df.R
\name{imp2df}
\alias{imp2df}
\title{Read impedance data from CHI760}
\usage{
imp2df(datafilename, wearea = NA)
}
\arguments{
\item{datafilename}{complete path to txt file}

\item{wearea}{WE area in sqcm, defaults to NA}
}
\value{
dataframe with data and metadata
}
\description{
This function converts CHI760 impedance data to dataframe while retaining metadata.
Reads: AC impedance, Impedance - Potential, and datafiles from related techniques
}

