% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsv2df.R
\name{lsv2df}
\alias{lsv2df}
\title{Read linear-scan voltammetry data from CHI760}
\usage{
lsv2df(datafilename, wearea = 1)
}
\arguments{
\item{datafilename}{text string with full path to experimental file}

\item{wearea}{(optional) area of working electrode (in sqcm)}
}
\value{
Dataframe with the following columns (and no extra attributes):
   $ sampleid        : chr (id)
   $ segment         : num (id)
   $ potential       : num (measure)
   $ current         : num (measure)
   $ charge          : num (measure)
   $ currentdensity  : num (measure)
   $ chargedensity   : num (measure)
   $ InitE           : num (id)
   $ FinalE          : num (id)
   $ ScanRate        : num (id)
   $ SampleInterval  : num (id)
   $ QuietTime       : num (id)
   $ Sensitivity     : num (id)
}
\description{
Reads LSV datafiles from CHI 760 potentiostat (potential, current, and charge)
and returns a dataframe with the data, the data attributes (experimental conditions),
and calculated current density and charge density.
}
\details{
The CH Instruments 760 potentiostat records all data using
   standard SI units, therefore this function assumes all potentials
   to be in volts, currents to be in amperes, charges in Coulombs,
   time in seconds, and so on.
}

