% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps2df.R
\name{mps2df}
\alias{mps2df}
\title{Read time vs current data from CHI760}
\usage{
mps2df(datafilename, wearea = 1)
}
\arguments{
\item{datafilename}{path to data file}

\item{wearea}{WE area in sqcm (defaults to 1)}
}
\value{
dataframe
}
\description{
Reads time vs current data (from CHI 760 potentiostat)
 and returns a dataframe with the data and the data attributes (experimental conditions).
}

