% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocp2df.R
\name{ocp2df}
\alias{ocp2df}
\title{Read time vs potential data from CHI760}
\usage{
ocp2df(datafilename)
}
\arguments{
\item{datafilename}{string with full path to experimental file
#'}
}
\value{
Dataframe with the following columns (and no extra attributes):
   $ sampleid        : chr
   $ time            : num
   $ potential       : num
   $ RunTime         : num
}
\description{
Reads time vs potential data from CHI760
}

